Update starting cluster offset in the scanning process

+++ src/FAT32Recovery.h
@@ -65,7 +65,7 @@ private:
 
     /*=============== Cluster and Sector Operations ===============*/
     // Read data from specified sector
-    bool readSector(uint32_t sector, void* buffer, uint32_t size);
+    bool readSector(uint64_t sector, void* buffer, uint64_t size);
     // Convert cluster number to sector number
     uint32_t clusterToSector(uint32_t cluster) const;
     // Get next cluster in the chain from FAT
